+++ src/test/kotlin/com/fwdekker/randomness/array/ArrayDecoratorTest.kt
@@ -7,6 +7,7 @@ import com.fwdekker.randomness.testhelpers.shouldValidateAsBundle
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.data.row
 import io.kotest.datatest.withData
+import io.kotest.matchers.collections.shouldHaveAtLeastSize
 import io.kotest.matchers.ints.shouldBeInRange
 import io.kotest.matchers.shouldBe
 
@@ -24,59 +25,69 @@ object ArrayDecoratorTest : FunSpec({
                 "returns default input if disabled" to
                     row(
                         ArrayDecorator(enabled = false, minCount = 3),
-                        "[i0]",
+                        "{i0}",
                     ),
                 "returns a single value" to
                     row(
                         ArrayDecorator(enabled = true, minCount = 1, maxCount = 1),
-                        "[[i0]]",
+                        "[{i0}]",
                     ),
                 "returns a fixed number of values" to
                     row(
                         ArrayDecorator(enabled = true, minCount = 3, maxCount = 3),
-                        "[[i0], [i1], [i2]]",
+                        "[{i0}, {i1}, {i2}]",
                     ),
                 "returns array with multi-char separator" to
                     row(
                         ArrayDecorator(enabled = true, separator = ";;"),
-                        "[[i0];;[i1];;[i2]]",
+                        "[{i0};;{i1};;{i2}]",
                     ),
                 "retains leading whitespace in separator" to
                     row(
                         ArrayDecorator(enabled = true, separator = ",  "),
-                        "[[i0],  [i1],  [i2]]",
+                        "[{i0},  {i1},  {i2}]",
                     ),
                 "converts escaped 'n' in separator to newline" to
                     row(
                         ArrayDecorator(enabled = true, separator = """\n"""),
-                        "[[i0]\n[i1]\n[i2]]",
+                        "[{i0}\n{i1}\n{i2}]",
                     ),
                 "applies affix decorator" to
                     row(
                         ArrayDecorator(enabled = true, affixDecorator = AffixDecorator(enabled = true, "(@)")),
-                        "([i0], [i1], [i2])",
+                        "({i0}, {i1}, {i2})",
                     ),
             )
         ) { (scheme, output) ->
-            scheme.generator = { count -> List(count) { "[i$it]" } }
+            scheme.generator = { count -> List(count) { "{i$it}" } }
 
             scheme.generateStrings()[0] shouldBe output
         }
 
-        test("generates the desired number of entries") {
+        test("generates the desired number of parts in each string") {
             val scheme = ArrayDecorator(enabled = true, minCount = 3, maxCount = 8)
-            scheme.generator = { count -> List(count) { "[i$it]" } }
+            scheme.generator = { count -> List(count) { "{i$it}" } }
 
-            scheme.generateStrings(count = 50).map { string -> string.count { it == ',' } + 1 }
+            scheme.generateStrings(count = 50)
+                .map { string -> string.count { it == ',' } + 1 }
                 .forEach { it shouldBeInRange 3..8 }
         }
 
-        test("appropriately chunks generator outputs") {
-            val scheme = ArrayDecorator(enabled = true)
-            var i = 0
-            scheme.generator = { count -> List(count) { "[i${i++}]" } }
+        test("generates an independently random number of parts per string") {
+            val scheme = ArrayDecorator(enabled = true, minCount = 1, maxCount = 8)
+            scheme.generator = { count -> List(count) { "{i$it}" } }
 
-            scheme.generateStrings(count = 2) shouldBe listOf("[[i0], [i1], [i2]]", "[[i3], [i4], [i5]]")
+            scheme.generateStrings(count = 50)
+                .map { string -> string.count { it == ',' } + 1 }
+                .distinct() shouldHaveAtLeastSize 2
+        }
+
+        test("appropriately splits parts into strings") {
+            val scheme = ArrayDecorator(enabled = true)
+            var partIdx = 0
+            scheme.generator = { count -> List(count) { "{i${partIdx++}}" } }
+
+            scheme.generateStrings(count = 2) shouldBe listOf("[{i0}, {i1}, {i2}]", "[{i3}, {i4}, {i5}]")
         }
     }
 
@@ -97,7 +108,7 @@ object ArrayDecoratorTest : FunSpec({
                     row(ArrayDecorator(affixDecorator = AffixDecorator(descriptor = """\""")), ""),
             )
         ) { (scheme, validation) ->
-            scheme.generator = { count -> List(count) { "[i$it]" } }
+            scheme.generator = { count -> List(count) { "{i$it}" } }
 
             scheme shouldValidateAsBundle validation
         }