+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## Unreleased
+## 3.1.0 -- 2023-12-08
 ### Added
 * Added ability to generate non-matching strings.
   ([#447](https://github.com/FWDekker/intellij-randomness/issues/447))
@@ -7,6 +7,9 @@
 ### Changed
 * When inserting arrays at multiple carets, the number of elements per array is now independently chosen for each array.
   ([#450](https://github.com/FWDekker/intellij-randomness/issues/450))
+
+### Deprecated
+* Minimum IDE version has been increased to 2023.1.
 
 
 ## 3.0.0 -- 2023-11-17